server/requirements.txt
earthengine-api
google-api-python-client
numpy
langgraph>=0.2.0
langchain-google-genai>=1.0.0
google-generativeai>=0.7.0
